+++ CHANGELOG.md
@@ -52,3 +52,4 @@
 -   Create specific fortnite auth client access token method to auth library.
 -   Kill multiple sessions method to auth library.
 -   Get external auths connections method to auth library.
+-   Unlink external auth method to auth library.